+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+- Addition of a `search` command. This allows you to run JQL either via the configuration file (you specify `--template` to find the correct template in the config file), or the CLI `--query` option. It displays the ID, and the title of the issues it finds.
+
 ## 2.1.0
 
 - Addition of the `-v` flag on the `sprint list` command.